Login y JWT para api. Tambien ajustes varios

+++ backend/src/main/java/com/grupo/backend/Controller/EducacionController.java
@@ -1,6 +1,9 @@
 package com.grupo.backend.Controller;
 
+import com.grupo.backend.Model.Educacion;
+import com.grupo.backend.Repository.EducacionRepository;
 import com.grupo.backend.Funciones;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,13 +17,12 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-
-import com.grupo.backend.Model.Educacion;
-import com.grupo.backend.Repository.EducacionRepository;
+import org.springframework.http.HttpHeaders;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 @Controller
 @RequestMapping(path = "/educacion")
-public class EducacionController {
+public class EducacionController extends ControllerGenerico {
     @Autowired
     private EducacionRepository educacionRepository;
 
@@ -32,7 +34,11 @@ public class EducacionController {
 
     @CrossOrigin(origins = "*")
     @PostMapping(path = "/add")
-    public @ResponseBody ResponseEntity<String> addEducacion(@RequestBody Educacion newEducacion) {
+    public @ResponseBody ResponseEntity<String> addEducacion(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody Educacion newEducacion) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<String>("Sin acceso", HttpStatus.UNAUTHORIZED);
+		}
+
         String error = this.validar(newEducacion);
 
         if (error != null) {
@@ -53,7 +59,11 @@ public class EducacionController {
 
     @CrossOrigin(origins = "*")
     @DeleteMapping(path = "/delete/{id}")
-    public ResponseEntity<HttpStatus> deleteEducacion(@PathVariable("id") int id) {
+    public ResponseEntity<HttpStatus> deleteEducacion(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable("id") int id) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
         try {
             educacionRepository.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -64,7 +74,11 @@ public class EducacionController {
 
     @CrossOrigin(origins = "*")
     @PutMapping("/update/{id}")
-    ResponseEntity<String> replaceEducacion(@RequestBody Educacion newEducacion, @PathVariable int id) {
+    ResponseEntity<String> replaceEducacion(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody Educacion newEducacion, @PathVariable int id) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<String>("Sin acceso", HttpStatus.UNAUTHORIZED);
+		}
+
         String error = this.validar(newEducacion);
 
         if (error != null) {
@@ -77,57 +91,54 @@ public class EducacionController {
                     educacion.setDescripcion(newEducacion.getDescripcion());
                     educacion.setDesde(newEducacion.getDesde());
                     educacion.setHasta(newEducacion.getHasta());
+                    educacionRepository.save(educacion);
 
                     return new ResponseEntity<String>("", HttpStatus.OK);
                 })
                 .orElseGet(() -> {
-                    return new ResponseEntity<String>("Error al actualizar educación",
-                            HttpStatus.INTERNAL_SERVER_ERROR);
+                    return new ResponseEntity<String>("Error al actualizar educación", HttpStatus.INTERNAL_SERVER_ERROR);
                 });
 
         return new ResponseEntity<String>("", HttpStatus.OK);
-
     }
 
     private String validar(Educacion newEducacion) {
-        String mensajeError = null;
-
         if (newEducacion.getTitulo().length() > 100) {
-            mensajeError = "Título excede la longitud permitida";
+            return "Título excede la longitud permitida";
         }
 
         if (newEducacion.getTitulo().length() == 0) {
-            mensajeError = "Título no puede estar vacío";
+            return "Título no puede estar vacío";
         }
 
         if (newEducacion.getDescripcion().length() > 250) {
-            mensajeError = "Descripción excede la longitud permitida";
+            return "Descripción excede la longitud permitida";
         }
 
         if (newEducacion.getDescripcion().length() == 0) {
-            mensajeError = "Descripción no puede estar vacía";
+            return "Descripción no puede estar vacía";
         }
 
         if (!Funciones.esNumero(newEducacion.getDesde())) {
-            mensajeError = "Desde tiene que ser un año";
+            return "Desde tiene que ser un año";
         }
 
         if (newEducacion.getDesde().length() == 0) {
-            mensajeError = "Desde no puede estar vacío";
+            return "Desde no puede estar vacío";
         }
 
         if (newEducacion.getHasta().compareTo(newEducacion.getDesde()) < 0) {
-            mensajeError = "Desde tiene que ser anterior a \"hasta\"";
+            return "Desde tiene que ser anterior a \"hasta\"";
         }
 
         if (!Funciones.esNumero(newEducacion.getHasta())) {
-            mensajeError = "Hasta tiene que ser un año";
+            return "Hasta tiene que ser un año";
         }
 
         if (newEducacion.getHasta().length() == 0) {
-            mensajeError = "Hasta no puede estar vacío";
+            return "Hasta no puede estar vacío";
         }
 
-        return mensajeError;
+        return null;
     }
 }
