+++ backend/src/main/java/com/grupo/backend/Controller/ExperienciaController.java
@@ -1,6 +1,9 @@
 package com.grupo.backend.Controller;
 
+import com.grupo.backend.Model.Experiencia;
+import com.grupo.backend.Repository.ExperienciaRepository;
 import com.grupo.backend.Funciones;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,13 +17,12 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-
-import com.grupo.backend.Model.Experiencia;
-import com.grupo.backend.Repository.ExperienciaRepository;
+import org.springframework.http.HttpHeaders;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 @Controller
 @RequestMapping(path = "/experiencia")
-public class ExperienciaController {
+public class ExperienciaController extends ControllerGenerico {
     @Autowired
     private ExperienciaRepository experienciaRepository;
 
@@ -32,7 +34,11 @@ public class ExperienciaController {
 
     @CrossOrigin(origins = "*")
     @PostMapping(path = "/add")
-    public @ResponseBody ResponseEntity<String> addExperiencia(@RequestBody Experiencia newExperiencia) {
+    public @ResponseBody ResponseEntity<String> addExperiencia(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody Experiencia newExperiencia) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<String>("Sin acceso", HttpStatus.UNAUTHORIZED);
+		}
+
         String error = this.validar(newExperiencia);
 
         if (error != null) {
@@ -55,7 +61,11 @@ public class ExperienciaController {
 
     @CrossOrigin(origins = "*")
     @DeleteMapping(path = "/delete/{id}")
-    public ResponseEntity<HttpStatus> deleteExperiencia(@PathVariable("id") int id) {
+    public ResponseEntity<HttpStatus> deleteExperiencia(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable("id") int id) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
         try {
             experienciaRepository.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -66,7 +76,11 @@ public class ExperienciaController {
 
     @CrossOrigin(origins = "*")
     @PutMapping("/update/{id}")
-    ResponseEntity<String> replaceExperiencia(@RequestBody Experiencia newExperiencia, @PathVariable int id) {
+    ResponseEntity<String> replaceExperiencia(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody Experiencia newExperiencia, @PathVariable int id) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<String>("Sin acceso", HttpStatus.UNAUTHORIZED);
+		}
+        
         String error = this.validar(newExperiencia);
 
         if (error != null) {
@@ -80,13 +94,12 @@ public class ExperienciaController {
                     experiencia.setDescripcion(newExperiencia.getDescripcion());
                     experiencia.setDesde(newExperiencia.getDesde());
                     experiencia.setHasta(newExperiencia.getHasta());
+                    experienciaRepository.save(experiencia);
 
                     return new ResponseEntity<String>("", HttpStatus.OK);
                 })
                 .orElseGet(() -> {
-                    return new ResponseEntity<String>("Error al actualizar experiencia",
-                            HttpStatus.INTERNAL_SERVER_ERROR);
-
+                    return new ResponseEntity<String>("Error al actualizar experiencia", HttpStatus.INTERNAL_SERVER_ERROR);
                 });
 
         return new ResponseEntity<String>("", HttpStatus.OK);
