+++ backend/src/main/java/com/grupo/backend/Controller/SkillController.java
@@ -2,13 +2,13 @@ package com.grupo.backend.Controller;
 
 import com.grupo.backend.Model.Skill;
 import com.grupo.backend.Repository.SkillRepository;
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,10 +16,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.http.HttpHeaders;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 @Controller
 @RequestMapping(path = "/skill")
-public class SkillController {
+public class SkillController extends ControllerGenerico {
     @Autowired
     private SkillRepository skillRepository;
 
@@ -31,7 +33,11 @@ public class SkillController {
 
     @CrossOrigin(origins = "*")
     @PostMapping(path = "/add")
-    public @ResponseBody ResponseEntity<String> addSkill(@RequestBody Skill newSkill) {
+    public @ResponseBody ResponseEntity<String> addSkill(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody Skill newSkill) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<String>("Sin acceso", HttpStatus.UNAUTHORIZED);
+		}
+
         String error = this.validar(newSkill);
 
         if (error != null) {
@@ -49,7 +55,11 @@ public class SkillController {
 
     @CrossOrigin(origins = "*")
     @DeleteMapping(path = "/delete/{id}")
-    public ResponseEntity<HttpStatus> deleteSkill(@PathVariable("id") int id) {
+    public ResponseEntity<HttpStatus> deleteSkill(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable("id") int id) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
         try {
             skillRepository.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -60,7 +70,11 @@ public class SkillController {
 
     @CrossOrigin(origins = "*")
     @PutMapping("/update/{id}")
-    ResponseEntity<String> updateSkill(@RequestBody Skill newSkill, @PathVariable int id) {
+    ResponseEntity<String> updateSkill(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody Skill newSkill, @PathVariable int id) {
+        if (!this.tokenValido(token)) {
+			return new ResponseEntity<String>("Sin acceso", HttpStatus.UNAUTHORIZED);
+		}
+
         String error = this.validar(newSkill);
 
         if (error != null) {
@@ -82,8 +96,6 @@ public class SkillController {
     }
 
     private String validar(Skill newSkill) {
-        String mensajeError = null;
-
         if (newSkill.getNivel() < 1) {
             return "El nivel tiene que ser mayor a 0";
         }
@@ -99,7 +111,6 @@ public class SkillController {
             return "Nombre no puede estar vacío";
         }
 
-        return mensajeError;
-
+        return null;
     }
 }
